pipe/pyproject.toml
[project]
name = "vibe-scanner"
version = "0.1.0"
description = "Antigravity Vibe Scanner Pipeline"
requires-python = ">=3.10"
dependencies = [
    "genai-processors",
    "google-genai",
    "python-dotenv",
    "aiohttp"
]

[tool.setuptools]
packages = ["pipe"]
